backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
# OLD: psycopg[binary]==3.2.1
psycopg[binary]>=3.2.10,<3.3
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.2
